requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.20
pydantic==1.10.7
requests==2.31.0

# Dev tools
pytest==7.3.1
black==23.3.0
mypy==1.3.0

# Tool integrations
python-jenkins==1.8.0
PyGithub==1.58.2
jfrog-python-sdk==1.0.0
